Fiks for sletting av avvent utbetaling perioder. Henter nå historikk for avvent perioder fra ubetalinger istedet for tilkjent ytelse, og forsøker ikke å sjekke om det er endring i avvent periode dersom det ikke dannes utbetaling for denne behandlingen.
