+++ app/src/main/kotlin/no/nav/aap/utbetal/tilkjentytelse/TilkjentYtelseRepository.kt
@@ -8,19 +8,7 @@ import no.nav.aap.komponenter.verdityper.GUnit
 import no.nav.aap.komponenter.verdityper.Prosent
 import no.nav.aap.utbetal.felles.YtelseDetaljer
 import no.nav.aap.utbetal.kodeverk.AvventÅrsak
-import java.time.LocalDate
-import java.time.LocalDateTime
 import java.util.*
-
-data class AvventPeriode (
-    val tilkjentYtelseId: Long,
-    val behandlingRef: UUID,
-    val avvent: Periode,
-    val overføres: LocalDate?,
-    val årsak: AvventÅrsak?,
-    val feilregistrering: Boolean,
-    val vedtakstidspunkt: LocalDateTime,
-)
 
 class TilkjentYtelseRepository(private val connection: DBConnection) {
 
@@ -49,42 +37,6 @@ class TilkjentYtelseRepository(private val connection: DBConnection) {
         lagreTilkjentYtelseTrekk(tilkjentYtelseId, tilkjentYtelse.trekk)
 
         return tilkjentYtelseId
-    }
-
-    fun hentTilkjentYtelseAvventHistorikk(saksnummer: Saksnummer): List<AvventPeriode> {
-        val sql = """
-            SELECT 
-                TILKJENT_YTELSE.ID AS TILKJENT_YTELSE_ID,
-                BEHANDLING_REF,
-                PERIODE,
-                OVERFORES,
-                ARSAK,
-                FEILREGISTRERING,
-                VEDTAKSTIDSPUNKT
-            FROM TILKJENT_YTELSE_AVVENT
-            JOIN TILKJENT_YTELSE ON TILKJENT_YTELSE.ID = TILKJENT_YTELSE_AVVENT.TILKJENT_YTELSE_ID
-            WHERE SAKSNUMMER = ?
-            ORDER BY VEDTAKSTIDSPUNKT
-        """.trimIndent()
-
-        return connection.queryList(sql) {
-            setParams {
-                setString(1, saksnummer.toString())
-            }
-
-            setRowMapper {
-                AvventPeriode(
-                    tilkjentYtelseId = it.getLong("tilkjent_ytelse_id"),
-                    behandlingRef = it.getUUID("BEHANDLING_REF"),
-                    avvent = it.getPeriode("PERIODE"),
-                    overføres = it.getLocalDateOrNull("OVERFORES"),
-                    årsak = AvventÅrsak.valueOf(it.getString("ARSAK")),
-                    feilregistrering = it.getBoolean("FEILREGISTRERING"),
-                    vedtakstidspunkt = it.getLocalDateTime("VEDTAKSTIDSPUNKT"),
-                )
-            }
-        }
-
     }
 
     private fun lagreTilkjentYtelsePerioder(tilkjentYtelseId: Long, tilkjentPerioder: List<TilkjentYtelsePeriode>) {
